+++ internal/db/migrations/00003_create_balances_table.go
@@ -14,8 +14,8 @@ func UpBalancesTable(ctx context.Context, tx *sql.Tx) error {
 	_, err := tx.ExecContext(ctx, `CREATE TABLE balances
 (
     uuid UUID PRIMARY KEY,
-    current INT DEFAULT 0,
-    withdrawn INT DEFAULT 0
+    current NUMERIC DEFAULT 0,
+    withdrawn NUMERIC DEFAULT 0
 );`)
 	return err
 }
